Remove bogus synthetic rows
All the context and explanation in the Agent PR linked to these changes.
The changes we should see here are the removal of rows with same PC as
the next row. These lines were synthetic markers inserted at the end of
a function boundary, and while some of them are correct, the removed
ones are wrong.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>