Remove redundant rows
As we binary search over the unwind information, any neighbouring rows
with the same unwind info can be summarised in just one.

This optimisation was obvious from early on and I added it back in
September last year (https://github.com/parca-dev/parca-agent/pull/794)
but it was removed later on.

Test Plan
========

See Agent PR.

Signed-off-by: Francisco Javier Honduvilla Coto <javierhonduco@gmail.com>